+++ index.html
@@ -55,7 +55,9 @@
       </ul>
       <p id="regUser" class="regUsers">Для добавления комментария<button class="reg-button">АВТОРИЗУЙТЕСЬ</button></p>
       <p id="load" class="load">...Подождите, список обновляется</p>
-      
+      <div id="input-comment">
+        
+      </div>
       <p id="loader" class="loader">...Добавляем комментарий</p>
     </div>
     <div id="reg"></div>
